// frontend/src/api/esquema.d.ts
export interface components {
    schemas: {
// ...
        LicitacaoResumo: {
            /** Id */
            id: number;
            /** Fonte */
            fonte: string;
            /** Numero */
            numero: string | null;
            /** Objeto */
            objeto: string | null;
            /** Modalidade */
            modalidade: string | null;
            /** Orgao */
            orgao: string | null;
            /** Abertura */
            abertura: string | null;
            /**
             * Situacao
             * @enum {string}
             */
            situacao: "aberta" | "andamento" | "suspensa" | "encerrada" | "cancelada" | "desconhecida";
            /** Rotulo */
            rotulo: string | null;
            /** Dias Para Abertura */
            dias_para_abertura: number | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}